#16 Get-contents endpoint
