+++ api/api.rest
@@ -9,3 +9,7 @@ Content-Type: application/json
 }
 
 ###
+
+GET http://localhost:9000/contents
+
+###
