+++ api/connect.js
@@ -1,0 +1,11 @@
+var mongoose = require('mongoose')
+
+mongoose.connect('mongodb://localhost/test', {useNewUrlParser: true,  useUnifiedTopology: true })
+
+var db = mongoose.connection;
+db.on('error', console.error.bind(console, 'connection error:'))
+db.once('open', function () {
+  console.log('we are connected')
+})
+
+module.exports = db
