--- api/routes/testAPI.js
@@ -1,10 +1,0 @@
-var express = require('express')
-
-var router = express.Router()
-
-router.get('/', function (req, res, next) {
-  res.send('API is working properly')
-})
-
-
-module.exports = router
