+++ api/server.js
@@ -1,55 +1,42 @@
-var createError = require('http-errors');
+// var createError = require('http-errors');
 var express = require('express');
 var path = require('path');
-var cookieParser = require('cookie-parser');
-var logger = require('morgan');
+// var cookieParser = require('cookie-parser');
+// var logger = require('morgan');
 var cors = require('cors');
+var db = require('./connect')
 
-var mongoose = require('mongoose')
-
-var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
-var testAPIRouter = require("./routes/testAPI");
+var contentRouter = require('./routes/content');
 
 var server = express();
 
 // view engine setup
-server.set('views', path.join(__dirname, 'views'));
-server.set('view engine', 'jade');
+// server.set('views', path.join(__dirname, 'views'));
+// server.set('view engine', 'jade');
 
 server.use(cors());
-server.use(logger('dev'));
+// server.use(logger('dev'));
 server.use(express.json());
-server.use(express.urlencoded({ extended: false }));
-server.use(cookieParser());
-server.use(express.static(path.join(__dirname, 'public')));
+// server.use(express.urlencoded({ extended: false }));
+// server.use(cookieParser());
+// server.use(express.static(path.join(__dirname, 'public')));
 
-server.use('/', indexRouter);
-server.use('/users', usersRouter);
-server.use('/testAPI', testAPIRouter);
+server.use('/', contentRouter);
 
-// catch 404 and forward to error handler
-server.use(function(req, res, next) {
-  next(createError(404));
-});
-
-// error handler
-server.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error');
-});
-
-mongoose.connect('mongodb://localhost/test', {useNewUrlParser: true,  useUnifiedTopology: true })
-
-var db = mongoose.connection;
-db.on('error', console.error.bind(console, 'connection error:'))
-db.once('open', function () {
-  console.log('we are connected')
-})
+// // catch 404 and forward to error handler
+// server.use(function(req, res, next) {
+//   next(createError(404));
+// });
+//
+// // error handler
+// server.use(function(err, req, res, next) {
+//   // set locals, only providing error in development
+//   res.locals.message = err.message;
+//   res.locals.error = req.app.get('env') === 'development' ? err : {};
+//
+//   // render the error page
+//   res.status(err.status || 500);
+//   res.render('error');
+// });
 
 module.exports = server;
